pyproject.toml
[build-system]
requires = ["maturin>=1.5,<2.0"]
build-backend = "maturin"

[project]
name = "gpm"
version = "0.1.0"
description = "Grasp Primary Module - Hardware interface via PyO3"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]
ui = [
    "flask>=3.0",
]

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "."
module-name = "gpm"
